Add rng provider trait
This commit defines an interface that allows an embedding
application to set an extension trait which can process
the functions that require a CryptoRng.

It does this without infecting the type signature of
essentially every part of sunset.

The implementation is 100% safe when building with the default getrandom
feature enabled, but when that feature is omitted and the new assignment
interface is added, there is some unsafe code to manage replacing the
pointer to that extension trait.

This is not likely totally sound in every scenario, but it happens
to work for me.

There's probably some cribbing that could be done from the log::logger()
and log::set_logger_racy() functions to make this more-safe.