Big changes: Fixed decoding SftpPacket
This changes the sftpmessge parameters to handle the three classes of SFTP Packets: Initialization, Requests and Responses.

The motivation is to add a mechanism to capture the Request Ids. Since only requests and responses contain that field but they do not include in their "inner" sftp Packets definitions (See Read, Write, etc) I thought that it was a good idea keeping them out of them.

To store them away from the Inner packets I decided adding extra parameters to the SftpPacket enum.

To do so the parameters have been reestructured so each class of sftp packet can be expanded with the required data.

As a side effect, the sftpmessages packets definition readability has been improved.

Certainly this adds complexity to the sftpmessages