WIP Added sftphandle.rs
This is the "traffic coordinator" for SFTP.
- Handles incoming packets deserialisation
- Handle the initialisation
- Calls SftpServer trait to handle client requests
- Handles incoming packets serialisation

The current implementation tries limiting its  dependencies to the minimum. Therefore it does not handle the SSH Channels itself but only buffers. This also allow the user to make decisions on the buffers.

the sftp demo has been updated to use this, delegating the sftp details to the sftphandle and demosftpserver