Simplifying SftpServer trait async and the Handle are out for now
I would like to keep those but since I want a structured of SftpServer to keep it's own internal state, I could not use async or a Handle with a trait definition. It is unfortunately unstable.

Still a WIP but now a client can PUT a small file and its content will be displayed in debug.

For now the buffers overflow on long SSH_FXP_WRITE packets and fail after trying to decode in the next loop iteration more of the same packet content.

The previous sftpserver trait has been stored in asyncsftpserver.rs