Adding basic obscuring of handler
This forces the SftpServer to use obscured handles that are size constrain.

It also introduces the handle manager, that is a helper that the SFTP server implementer can use for dealing with the obscured handle and a private handle, that can contain all the information that the local server requires.