WIP: Refactoring opaque_file_handle -> obscure_file_handle and adding traits
Will not compile ATM

First, I am still working on the details to make the use of traits to define the DemoSftpServer flexible. In this commit there is too much rigidity in the  DemoSftpServer which does not allow my intended vision.

Second, there is an unrelated refactoring here where SFTPSink and SftpSource have been extracted from sftphandle.rs to their own files for clarity