[skip ci] WIP: Some issues in DemoSftpServer DirEntriesCollection but ready to make the SftpServer trait functions async
The length calculation seems erroneous but I am more worried about the async trait

[skip ci] WIP: Progress with readdir. Needs tyding up Issue with pipe

Looks like the number of sent items in the pipe does not match received items. I am going to add a send/receive total bytes counter. It might also be the cause of the write errors

[skip ci] WIP: Cargo lock update long forgotten. Apologies