[skip ci] WIP: First ls command successful
Not ready yet. In order to answer  EOF to a SSH_SXP_READDIR after all the dir items have been already sent. I am storing a `sftpserver::ReadStatus` passed as a result on  `SftpServer.dirread()` with PendingData or EndOfFile.

This would not work well in scenarios where there is more than one read transaction in course.

I have decided that I am going to   make the SftpServer implementators the responsibility to send the EOF status message but I will help them by providing a detailed trait documentation.

To test this listing you may run the demo/sftp/std and run the script testing/test_read_dir.sh

mind that the hard link counter is not recovered as it is not part of SFTP V3 implementation