async: Fix discarded channel input data
Channel input was inadvertently being discarded because of an incorrect
refcount comparison (wrong order of operations).

This would cause data to be discarded if async progress() runs before
the ChanIn has consumed data. In sunsetc this never seems to occur,
maybe due to specifics how the Futures are executed by
embassy_futures::select?

In work in progress SFTP's demo, incoming data seems to be
intermittently discarded.

Fixes: f06f7cea77b8 ("Move channel wakers into core sunset crate")