[skip ci] Exploring logs: Found important issue in OutputChannelHandler
Not all the data that the producer tries to send reach the consumer, that is the one that writes the data in the ChannelOut.

This is likely to be the responsible for the stuck communication.

Apart from that, I added more testing and fixed issue in Proto::Data (SSH_FXP_DATA). It had too many fields.

More details to come